Implemented Facebook login with rudimentary firebase, making the maing App functionality now with user accounts and languages
