+++ package.json
@@ -9,10 +9,14 @@
   },
   "dependencies": {
     "expo": "^32.0.0",
+    "firebase": "^5.11.1",
+    "native-base": "^2.12.1",
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
     "react-native-audio": "^4.3.0",
     "react-native-fetch-blob": "^0.10.8",
+    "react-native-firebase": "^5.3.1",
+    "react-native-gifted-chat": "^0.8.0",
     "rn-fetch-blob": "^0.10.15",
     "socket.io-client": "^2.2.0"
   },
